Remove unnecessary casts

+++ algo/src/main/java/org/neo4j/graphalgo/impl/labelprop/LabelPropagation.java
@@ -151,7 +151,7 @@ public class LabelPropagation extends Algorithm<LabelPropagation> {
 
     private List<StepRunner> stepRunners(Direction direction) {
         long nodeCount = graph.nodeCount();
-        long batchSize = ParallelUtil.adjustedBatchSize(nodeCount, (long) this.batchSize);
+        long batchSize = ParallelUtil.adjustedBatchSize(nodeCount, this.batchSize);
 
         Collection<PrimitiveLongIterable> nodeBatches = LazyBatchCollection.of(
                 nodeCount,
